remove cnm code
