--- cnm/config.json
@@ -1,33 +1,0 @@
-{
-	"description": "Azure VNET plugin",
-	"documentation": "https://github.com/Azure/azure-container-networking/",
-	"entrypoint": ["/usr/bin/azure-vnet-plugin"],
-	"interface": {
-		"types": ["docker.networkdriver/1.0", "docker.ipamdriver/1.0"],
-		"socket": "azure-vnet.sock"
-	},
-	"network": {
-		"type": "host"
-	},
-	"mounts": [
-		{
-			"name": "logs",
-			"description": "Mount /var/log to expose plugin logs to host",
-			"source": "/var/log",
-			"destination": "/var/log",
-			"type": "bind",
-			"options": ["rbind", "rw"]
-		},
-		{
-			"name": "modules",
-			"description": "Mount /lib/modules to load ebtables kernel module on demand",
-			"source": "/lib/modules",
-			"destination": "/lib/modules",
-			"type": "bind",
-			"options": ["rbind", "r"]
-		}
-	],
-	"linux": {
-		"capabilities": ["CAP_SYS_ADMIN", "CAP_NET_ADMIN", "CAP_SYS_MODULE"]
-	}
-}
